Split out the code that uses a set of bindings to serialize to/from the instant execution cache from the code that builds the bindings. This allows the code that uses the bindings to be reused in a context where a different set of bindings is required, specifically when serializing transform nodes.
